Use CollaborationContext
