+++ frontend/src/presentation/routes/AppRoutes.tsx
@@ -11,6 +11,7 @@ import { ProtectedRoute } from "presentation/pages/ProtectedRoute";
 import { AdminProtectedRoute } from "presentation/pages/AdminProtectedRoute";
 import { MatchmakingProvider } from "domain/context/MatchmakingContext";
 import CollaborationRoomPage from "presentation/pages/CollaborationRoomPage";
+import { CollaborationProvider } from "domain/context/CollaborationContext";
 
 const AppRoutes: React.FC = () => {
     return (
@@ -22,7 +23,14 @@ const AppRoutes: React.FC = () => {
                     <Route path="/room/:roomId/:matchUserId" element={<QuestionPage />} />
 
                     <Route element={<ProtectedRoute />}>
-                        <Route path="/room/:roomId" element={<CollaborationRoomPage />} />
+                        <Route
+                            path="/room/:roomId"
+                            element={
+                                <CollaborationProvider>
+                                    <CollaborationRoomPage />
+                                </CollaborationProvider>
+                            }
+                        />
                         <Route path="/" element={<HomePage />} />
 
                         <Route element={<AdminProtectedRoute />}>
